Clean code. Refactored namespace Strategy to StrategyPatternExample.
