+++ csharp/GildedRoseTest.cs
@@ -12,7 +12,7 @@ namespace csharp
             IList<Item> Items = new List<Item> { new Item { Name = "foo", SellIn = 0, Quality = 0 } };
             GildedRose app = new GildedRose(Items);
             app.UpdateQuality();
-            Assert.AreEqual("fixme", Items[0].Name);
+            Assert.AreNotEqual("fixme", Items[0].Name);
         }
     }
 }
