+++ csharp/StrategyPatternExample/ICategoryStrategy.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace csharp.Strategy
+namespace csharp.StrategyPatternExample
 {
     interface ICategoryStrategy
     {
