+++ csharp/StrategyPatternExample/ItemWrapperContext.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace csharp.Strategy
+namespace csharp.StrategyPatternExample
 {
     public class ItemWrapperContext
     {
