+++ csharp/StrategyPatternExample/NextExpiredImproveQualityStrategy.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace csharp.Strategy
+namespace csharp.StrategyPatternExample
 {
     class NextExpiredImproveQualityStrategy : ICategoryStrategy
     {
